notebooks/requirements.txt
numpy==1.24.4
pandas==2.2.1
peft==0.15.0
pillow==10.4.0
pillow_heif==0.22.0
pytorch-lightning==2.5.1
PyYAML==6.0.1
scipy==1.13.1
tensorboard==2.9.0
timm==1.0.15
tqdm==4.66.4
transformers==4.50.0
wandb==0.19.8
qdrant-client==1.12.1
open-metric-learning==4.0.0
omegaconf==2.3.0
hydra-core==1.3.2
qdrant-client==1.12.0
pydantic==2.11.7